# BookBackend/api/openAPI.yaml
openapi: "3.0.3"
info:
  title: "BookBackend API"
  description: "BookBackend API"
  version: "1.0.0"
servers:
  - url: "https://BookBackend"
paths:
...
  /cd:
    get:
      tags:
        - "CD"
      summary: "GET cd"
      operationId: "getAllCD"
      responses:
        "200":
          $ref: "#/components/responses/CdList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
        tags:
            - "CD"
        summary: "PUT cd"
        operationId: "CreateCd"
        requestBody:
            $ref: "#/components/requestBodies/CreateCd"
        responses:
            "200":
              $ref: "#/components/responses/Cd"
            "400":
              $ref: "#/components/responses/BadRequest"
            "401":
              $ref: "#/components/responses/Unauthorized"
        security:
            - prod: [ ]
  /cd/{cdId}:
    get:
      tags:
        - "CD"
      summary: "GET cd/{id}"
      operationId: "getCdById"
      parameters:
        - $ref: "#/components/parameters/cdId"
      responses:
        "200":
          $ref: "#/components/responses/Cd"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
        tags:
            - "CD"
        summary: "PUT cd/{id}"
        operationId: "updateCd"
        parameters:
            - $ref: "#/components/parameters/cdId"
        requestBody:
            $ref: "#/components/requestBodies/UpdateCd"
        responses:
            "200":
              $ref: "#/components/responses/Cd"
            "400":
              $ref: "#/components/responses/BadRequest"
            "401":
              $ref: "#/components/responses/Unauthorized"
        security:
            - prod: [ ]
    delete:
        tags:
            - "CD"
        summary: "DELETE cd/{id}"
        operationId: "deleteCD"
        parameters:
            - $ref: "#/components/parameters/cdId"
        responses:
            "200":
              description: "OK"
            "401":
              $ref: "#/components/responses/Unauthorized"
        security:
            - prod: [ ]
...
components:
  parameters:
...
    cdId:
      name: cdId
      in: path
      description: ID of cd to return
      required: true
      schema:
        type: integer
        format: int64
  requestBodies:
    CreateCd:
      description: "Create cd"
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CreateCd"
    UpdateCd:
        description: "Update cd"
        required: true
        content:
            application/json:
              schema:
                  $ref: "#/components/schemas/UpdateCd"
...
  responses:
...
    Cd:
      description: "OK"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Cd"
...
    BadRequest:
        description: "Bad Request"
        content:
            application/json:
              schema:
                  $ref: "#/components/schemas/Error"
    Unauthorized:
        description: "Unauthorized"
        content:
            application/json:
              schema:
                  $ref: "#/components/schemas/Error"
  schemas:
...
    Cd:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1234
        title:
          type: "string"
          example: "The Lord of the Rings"
        discription:
          type: "string"
          example: "Der erste Band der Trilogie"
        artist:
          type: "string"
          example: "J.R.R. Tolkien"
        label:
          type: "string"
          example: "Hörbuch"
    UpdateCd:
      type: "object"
      properties:
        title:
          type: "string"
          example: "The Lord of the Rings"
        discription:
          type: "string"
          example: "Der erste Band der Trilogie"
        artist:
          type: "string"
          example: "J.R.R. Tolkien"
        publisher:
          type: "string"
          example: "xyz-publisher"
        label:
          type: "string"
          example: "Hörbuch"
    CreateCd:
      type: "object"
      properties:
        title:
          type: "string"
          example: "The Lord of the Rings"
        discription:
          type: "string"
          example: "Der erste Band der Trilogie"
        artist:
          type: "string"
          example: "J.R.R. Tolkien"
        publisher:
          type: "string"
          example: "xyz-publisher"
        label:
          type: "string"
          example: "Hörbuch"
...
    Error:
        type: "object"
        properties:
            message:
              type: "string"
              example: "Bad Request"
  securitySchemes:
    prod:
      scheme: bearer
      type: http
      bearerFormat: JWT